+++ orchestrator/distlock/__init__.py
@@ -55,9 +55,7 @@ distlock_manager = cast(DistLockManager, wrapped_distlock_manager)
 # The Global DistLockManager is set after calling this function
 def init_distlock_manager(settings: AppSettings) -> DistLockManager:
     wrapped_distlock_manager.update(
-        DistLockManager(
-            settings.ENABLE_DISTLOCK_MANAGER, settings.DISTLOCK_BACKEND, (settings.CACHE_HOST, settings.CACHE_PORT)
-        )
+        DistLockManager(settings.ENABLE_DISTLOCK_MANAGER, settings.DISTLOCK_BACKEND, settings.CACHE_URI)
     )
     return distlock_manager