+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
-## [Unreleased]
+## [1.1.0rc1]
+
+### Breaking change
+ - Changed `settings.CACHE_HOST` and `settings.CACHE_PORT` to `settings.CACHE_URI` 
+
+## [1.0.2]
 
 - Allow user to extend the default translation set, instead of overwriting everything
-- Return worker status information in the `/api/settings/status` endpoint
+- Return worker status information in the `/api/settings/worker-status` endpoint
 
 ## [0.4.0-rc1](https://github.com/workfloworchestrator/orchestrator-core/tree/0.4.0-rc1) (2022-03-08)